+++ netcdf/netcdf_profile.cc
@@ -154,7 +154,7 @@ void netcdf_profile::fill_missing() {
     const int alt_num = this->_axis[0]->size() ;
     const int lat_num = this->_axis[1]->size() ;
     const int lng_num = this->_axis[2]->size() ;
-    unsigned index[3] ;
+    size_t index[3] ;
 
     // compute average value at each depth
 
@@ -185,7 +185,7 @@ void netcdf_profile::fill_missing() {
             if ( index[0] <= 0 ) {
                 average.data( index, NAN ) ;
             } else {
-                unsigned prev_index[1] ;
+                size_t prev_index[1] ;
                 prev_index[0] = index[0] - 1 ;
                 average.data( index, average.data(prev_index) ) ;
             }
