+++ waveq3d/spreading_hybrid_gaussian.h
@@ -159,7 +159,7 @@ class USML_DECLSPEC spreading_hybrid_gaussian : public spreading_model {
      * @return              Intensity of ray at this point.
      */
     virtual const vector<double>& intensity(
-        const wposition1& location, unsigned de, unsigned az,
+        const wposition1& location, size_t de, size_t az,
         const vector<double>& offset, const vector<double>& distance ) ;
 
     /**
@@ -176,7 +176,7 @@ class USML_DECLSPEC spreading_hybrid_gaussian : public spreading_model {
      * @return              Half-width of cell in the DE direction.
      */
     virtual double width_de(
-        unsigned de, unsigned az, const vector<double>& offset ) ;
+        size_t de, size_t az, const vector<double>& offset ) ;
 
     /**
      * Interpolate the half-width of a cell in the AZ direction.
@@ -192,7 +192,7 @@ class USML_DECLSPEC spreading_hybrid_gaussian : public spreading_model {
      * @return              Half-width of cell in the AZ direction.
      */
     virtual double width_az(
-        unsigned de, unsigned az, const vector<double>& offset ) ;
+        size_t de, size_t az, const vector<double>& offset ) ;
 
   private:
 
@@ -207,7 +207,7 @@ class USML_DECLSPEC spreading_hybrid_gaussian : public spreading_model {
      * @param  distance     Offsets in distance units.
      * @return              Intensity of ray at this point.
      */
-    void intensity_de( unsigned de, unsigned az,
+    void intensity_de( size_t de, size_t az,
         const vector<double>& offset, const vector<double>& distance ) ;
 
     /**
@@ -221,7 +221,7 @@ class USML_DECLSPEC spreading_hybrid_gaussian : public spreading_model {
      * @param  distance     Offsets in distance units.
      * @return              Intensity of ray at this point.
      */
-    void intensity_az( unsigned de, unsigned az,
+    void intensity_az( size_t de, size_t az,
         const vector<double>& offset, const vector<double>& distance ) ;
 
 } ;
